altering time-Variables from 'unsigned int' to 'size_t' without much thinking,in order to make this branch compilable - I didn't apply any tests
git-svn-id: http://svn.code.sf.net/p/sumo/code/branches/colombo@13946 29f149b2-0c63-11e1-8f5c-a347f6a9acd0

+++ src/microsim/traffic_lights/MSSOTLTrafficLightLogic.h
@@ -117,7 +117,8 @@ protected:
 	 * This member has to contain the switching logic for SOTL policies
 	 */
 	
-	virtual unsigned int decideNextPhase();
+	//	virtual unsigned int decideNextPhase();
+	virtual size_t decideNextPhase();
 
 
 	virtual bool canRelease() = 0;
@@ -223,7 +224,7 @@ private:
 	* To reset the cars-timesteps counter when a target phase is newly selected
 	* If phaseStep is not a target phase nothing happens
 	*/
-	void resetCTS(unsigned int phaseStep);
+	void resetCTS(size_t phaseStep);
 
 };
